changed docker name
