--- services/apiserver/cmd/main.go
@@ -1,12 +1,0 @@
-package main
-
-import (
-	"context"
-
-	apiserver "github.com/solo-io/service-mesh-hub/services/apiserver/pkg"
-)
-
-func main() {
-	ctx := context.Background()
-	apiserver.Run(ctx)
-}
