+++ services/mesh-apiserver/api/v1/services.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 package grpc.solo.io;
-option go_package = "github.com/solo-io/service-mesh-hub/services/apiserver/pkg/api/v1";
+option go_package = "github.com/solo-io/service-mesh-hub/services/mesh-apiserver/pkg/api/v1";
 
 import "service-mesh-hub/api/discovery/v1alpha1/mesh_service.proto";
 import "service-mesh-hub/api/core/v1alpha1/ref.proto";
