+++ services/mesh-apiserver/pkg/server/server.go
@@ -12,8 +12,8 @@ import (
 	"github.com/rotisserie/eris"
 	"github.com/solo-io/go-utils/contextutils"
 	"github.com/solo-io/service-mesh-hub/pkg/env"
-	rpc_v1 "github.com/solo-io/service-mesh-hub/services/apiserver/pkg/api/v1"
-	"github.com/solo-io/service-mesh-hub/services/apiserver/pkg/server/health_check"
+	rpc_v1 "github.com/solo-io/service-mesh-hub/services/mesh-apiserver/pkg/api/v1"
+	"github.com/solo-io/service-mesh-hub/services/mesh-apiserver/pkg/server/health_check"
 	"go.opencensus.io/plugin/ocgrpc"
 	"go.opencensus.io/stats/view"
 	"google.golang.org/grpc"
